Backend/requirements.txt
fastapi
uvicorn
pyserial
websockets
pydantic
    